Refactor: eggday types
